+++ admin/backend/src/main/java/com/qusu/mybiomni/dao/mysql/model/MessageDOExample.java
@@ -205,52 +205,52 @@ public class MessageDOExample {
             return (Criteria) this;
         }
 
-        public Criteria andIdEqualTo(Long value) {
+        public Criteria andIdEqualTo(Integer value) {
             addCriterion("id =", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotEqualTo(Long value) {
+        public Criteria andIdNotEqualTo(Integer value) {
             addCriterion("id <>", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThan(Long value) {
+        public Criteria andIdGreaterThan(Integer value) {
             addCriterion("id >", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThanOrEqualTo(Long value) {
+        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
             addCriterion("id >=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThan(Long value) {
+        public Criteria andIdLessThan(Integer value) {
             addCriterion("id <", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThanOrEqualTo(Long value) {
+        public Criteria andIdLessThanOrEqualTo(Integer value) {
             addCriterion("id <=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdIn(List<Long> values) {
+        public Criteria andIdIn(List<Integer> values) {
             addCriterion("id in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotIn(List<Long> values) {
+        public Criteria andIdNotIn(List<Integer> values) {
             addCriterion("id not in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdBetween(Long value1, Long value2) {
+        public Criteria andIdBetween(Integer value1, Integer value2) {
             addCriterion("id between", value1, value2, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotBetween(Long value1, Long value2) {
+        public Criteria andIdNotBetween(Integer value1, Integer value2) {
             addCriterion("id not between", value1, value2, "id");
             return (Criteria) this;
         }
@@ -265,52 +265,52 @@ public class MessageDOExample {
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdEqualTo(Long value) {
+        public Criteria andConversationIdEqualTo(Integer value) {
             addCriterion("conversation_id =", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdNotEqualTo(Long value) {
+        public Criteria andConversationIdNotEqualTo(Integer value) {
             addCriterion("conversation_id <>", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdGreaterThan(Long value) {
+        public Criteria andConversationIdGreaterThan(Integer value) {
             addCriterion("conversation_id >", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdGreaterThanOrEqualTo(Long value) {
+        public Criteria andConversationIdGreaterThanOrEqualTo(Integer value) {
             addCriterion("conversation_id >=", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdLessThan(Long value) {
+        public Criteria andConversationIdLessThan(Integer value) {
             addCriterion("conversation_id <", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdLessThanOrEqualTo(Long value) {
+        public Criteria andConversationIdLessThanOrEqualTo(Integer value) {
             addCriterion("conversation_id <=", value, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdIn(List<Long> values) {
+        public Criteria andConversationIdIn(List<Integer> values) {
             addCriterion("conversation_id in", values, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdNotIn(List<Long> values) {
+        public Criteria andConversationIdNotIn(List<Integer> values) {
             addCriterion("conversation_id not in", values, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdBetween(Long value1, Long value2) {
+        public Criteria andConversationIdBetween(Integer value1, Integer value2) {
             addCriterion("conversation_id between", value1, value2, "conversationId");
             return (Criteria) this;
         }
 
-        public Criteria andConversationIdNotBetween(Long value1, Long value2) {
+        public Criteria andConversationIdNotBetween(Integer value1, Integer value2) {
             addCriterion("conversation_id not between", value1, value2, "conversationId");
             return (Criteria) this;
         }
